+++ src/models/ComponentAttribute.js
@@ -37,7 +37,7 @@ class ComponentAttribute {
      *
      * @type {(string | boolean | number | Array | ComponentAttribute[])}
      */
-    this.value = value || null;
+    this.value = value ?? null;
     /**
      * The type of the attribute, reel type of the attribute, to check if it's match with the
      * ComponentAttributeDefinition.
@@ -52,6 +52,99 @@ class ComponentAttribute {
      */
     this.definition = definition || null;
   }
-}
 
+  /**
+   * Check if the attribute has an error.
+   *
+   * @returns {boolean} - true if an error occurs otherwise false.
+   */
+  hasError() {
+    if (this.definition === null) {
+      if (this.value === null) {
+        return true;
+      }
+
+      if (this.type.toLowerCase() === 'array') {
+        return !Array.isArray(this.value);
+      }
+
+      // eslint-disable-next-line valid-typeof
+      return typeof this.value !== this.type.toLowerCase();
+    }
+
+    return this.__typeOfValueValidation()
+      || this.__ruleValueValidation()
+      || this.__ruleMinAndMaxValidation()
+      || this.__ruleRegexValidation();
+  }
+
+  /**
+   * Check if the attribute type matches one of the JavaScript types.
+   *
+   * @returns {boolean} - true if the type is not good otherwise false.
+   * @private
+   */
+  __typeOfValueValidation() {
+    const type = this.type.toLowerCase();
+
+    if (type === 'array') {
+      if (this.definition.type === 'Link') {
+        return !Array.isArray(this.value) || this.value.some((link) => typeof link !== 'string');
+      }
+
+      return !Array.isArray(this.value);
+    }
+
+    if (type === 'object') {
+      return Object.prototype.toString.call(this.value) !== '[object Object]';
+    }
+
+    // eslint-disable-next-line valid-typeof
+    return typeof this.value !== type;
+  }
+
+  /**
+   * Check if values rule include this attribute value.
+   *
+   * @returns {boolean} - true if values rule doesn't include this attribute value otherwise false.
+   * @private
+   */
+  __ruleValueValidation() {
+    return this.definition.rules.values
+      && !this.definition.rules.values.includes(this.value);
+  }
+
+  /**
+   * Check if the value or its length follows min and max rules.
+   *
+   * @returns {boolean} - true if the value is smaller than the "min" rule
+   * or bigger than the "max" rule otherwise false.
+   * @private
+   */
+  __ruleMinAndMaxValidation() {
+    const value = typeof this.value === 'number' ? this.value : this.value.length;
+
+    if (this.definition.rules.min !== null && value < this.definition.rules.min) {
+      return true;
+    }
+
+    return this.definition.rules.max !== null && value > this.definition.rules.max;
+  }
+
+  /**
+   * Check if the value matches the regex rule.
+   *
+   * @returns {boolean} - true if attribute's value doesn't match the regex rule otherwise false.
+   * @private
+   */
+  __ruleRegexValidation() {
+    if (this.definition.rules.regex !== null) {
+      const regex = new RegExp(this.definition.rules.regex);
+
+      return !regex.test(this.value);
+    }
+
+    return false;
+  }
+}
 export default ComponentAttribute;