+++ changelog.md
@@ -10,14 +10,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add a lot of e2e tests to improve e2e coverage
-- Add `getDragTarget` method to `DefaultDrawer.js`.
 - Add an id to the component's icons for cypress tests.
+- Add `hasError` method to `Component.js` to check if a component has an error
+- Add methods checking error for all types to `ComponentAttribute.js`
+- Add error icon on component SVGs.
 
 ### Changed
 
 - Replace esdoc by jsdoc
 - Make configuration readable by user in the configuration file.
 - Save component's positions after the layout is automatically generated.
+- Change `ComponentAttributeDefinition.` containerRef's type from array to string.
 
 ### Fixed